Actualizacion de todo jsjs

+++ SumaDeVectores/interfaz.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
-extern void suma_vectores(float *a, float *b, float *res);
+extern void suma_vectores(double *a, double *b, double *res);
 
 int main(){
-    float vecA[]= {1.1f,2.2f,3.3f,4.4f};
-    float vecB[]= {10.0f,20.0f,30.0f,40.0f};
-    float resultado[4];
+    double vecA[]= {1.1,2.2,3.3,4.4};
+    double vecB[]= {10.0,20.0,30.0,40.0};
+    double resultado[4];
 
     suma_vectores(vecA,vecB,resultado);
 
-    printf("Resultado SIMD: [%.1f, %.1f, %.1f, %.1f]\n", resultado[0], resultado[1], resultado[2], resultado[3]);
+    printf("Resultado SIMD: [%.1lf, %.1lf, %.1lf, %.1lf]\n", resultado[0], resultado[1], resultado[2], resultado[3]);
 
     return 0;
 }
